requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
commonmark==0.9.1
discord.py==1.7.3
idna==3.3
jarowinkler==1.0.1
multidict==6.0.2
mutagen==1.45.1
numpy==1.22.3
opuslib==3.0.1
pycparser==2.21
pycryptodomex==3.14.1
Pygments==2.11.2
PyNaCl==1.4.0
pytube==12.0.0
rapidfuzz==2.0.6
requests==2.27.1
rich==12.0.0
six==1.16.0
soupsieve==2.3.1
spotdl==3.9.3
spotipy==2.19.0
typing_extensions==4.1.1
Unidecode==1.3.4
urllib3==1.26.8
websockets==10.2
yarl==1.7.2
youtube-dl==2021.12.17
yt-dlp==2022.3.8.2
ytmusicapi==0.21.0
